requirements.txt
asgiref==3.11.0
Django==5.0.14
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.5
